+++ imio/dms/mail/adapters.py
@@ -1776,13 +1776,13 @@ class OMApprovalAdapter(object):
                 )
                 helper_view.pod_template = download_template.UID()
                 helper_view.output_format = "pdf"
+                # QR code, with borders
                 gen_context = {
                     "context": self.context,
                     "portal": api.portal.get(),
                     "view": helper_view,
-                    "download_barcode": generate_barcode(
-                        download_url, scale=1, extra_args=["--secure=10", "--noquietzones"]
-                    ).read(),
+                    "download_barcode": generate_barcode(download_url, barcode=58, scale=3, resize=0.5,
+                                                         extra_args=["--vwhitesp=2", "--whitesp=2"]).read(),
                     "download_url": download_url,
                     "max_download_date": get_max_download_date(None, adate=datetime.date.today()),
                     "render_download_barcode": True,
@@ -1838,9 +1838,8 @@ class OMApprovalAdapter(object):
             update_dict_with_validation(
                 gen_context,
                 {
-                    "download_barcode": generate_barcode(
-                        download_url, scale=1, extra_args=["--secure=10", "--noquietzones"]
-                    ).read(),
+                    "download_barcode": generate_barcode(download_url, barcode=58, scale=3, resize=0.5,
+                                                         extra_args=["--vwhitesp=2", "--whitesp=2"]).read(),
                     "download_url": download_url,
                     "max_download_date": get_max_download_date(None, adate=datetime.date.today()),
                     "render_download_barcode": True,